xenforo: Fix xenforo cache breaking
This is the thing where if we try to fetch a chapter and we get a value from
the cache and the chapter isn't there that we go refetch from the source.

This is necessary when a new chapter is posted to the same forum page as a
previous chapter.