ClearLine on error exit, and clarify failure to fetch.
The progress bar had previously been left on the screen (it looks like
it resets the cursor position to the begining of the line). This
results in the error message overwriting it giving errors like

  An error occured: Cannot read property 'replace' of \
  nullwn.org/works/1234567890

In addition, this change also throws a new Error when the url results
in an empty fic object.