*: Add code for loading author data (incomplete)
This adds code that can scrape author pages and merge them across sites,
storing them in a central .authors.json file.  The UX around all of this is
quite a mess right now, just the bare minimum for my needs.  Nonetheless it
sits independent so it's going out and I'll revisit it later before giving
it docs and a public face.