get-author: Ignore known archive authors (FIXME)
This potential list is arbitrarily long.  These authors are used when
importing an archive into AO3 when the original archive author can't be
correlated with an existing AO3 author.  Pruning them is important or they
end up linking together unrelated authors.  (Ironically, we are often able
to correlate the authors with accounts even when AO3 could not, due to our
matching with FFNet.)